+++ lib/active_harness/request/models.rb
@@ -134,7 +134,7 @@ module ActiveHarness
       @models = []
     end
 
-    def use(provider:, model:, temperature: nil, name: nil, size: nil, quality: nil, retry_attempts: nil, retry_delay: nil)
+    def use(provider:, model:, temperature: nil, name: nil, size: nil, quality: nil, retry_attempts: nil, retry_delay: nil, questions: nil)
       @models << {
         provider:       provider,
         model:          model,
@@ -143,7 +143,8 @@ module ActiveHarness
         size:           size,
         quality:        quality,
         retry_attempts: retry_attempts,
-        retry_delay:    retry_delay
+        retry_delay:    retry_delay,
+        questions:      questions
       }.compact
     end